Move exception code into cp0 module.
