+++ src/device/r4300/cp0.c
@@ -23,7 +23,6 @@
 #include <string.h>
 
 #include "cp0.h"
-#include "exception.h"
 #include "new_dynarec/new_dynarec.h" /* for NEW_DYNAREC_ARM */
 #include "recomp.h"
 
@@ -141,5 +140,142 @@ void cp0_update_count(struct r4300_core* r4300)
    if (g_DebuggerActive && !r4300->delay_slot) update_debugger(*r4300_pc(r4300));
 #endif
 */
+}
+
+void TLB_refill_exception(struct r4300_core* r4300, uint32_t address, int w)
+{
+    uint32_t* cp0_regs = r4300_cp0_regs(&r4300->cp0);
+    int usual_handler = 0, i;
+
+    if (r4300->emumode != EMUMODE_DYNAREC && w != 2) {
+        cp0_update_count(r4300);
+    }
+
+    cp0_regs[CP0_CAUSE_REG] = (w == 1)
+        ? CP0_CAUSE_EXCCODE_TLBS
+        : CP0_CAUSE_EXCCODE_TLBL;
+
+    cp0_regs[CP0_BADVADDR_REG] = address;
+    cp0_regs[CP0_CONTEXT_REG] = (cp0_regs[CP0_CONTEXT_REG] & UINT32_C(0xFF80000F))
+        | ((address >> 9) & UINT32_C(0x007FFFF0));
+    cp0_regs[CP0_ENTRYHI_REG] = address & UINT32_C(0xFFFFE000);
+
+    if (cp0_regs[CP0_STATUS_REG] & CP0_STATUS_EXL)
+    {
+        generic_jump_to(r4300, UINT32_C(0x80000180));
+
+
+        if (r4300->delay_slot == 1 || r4300->delay_slot == 3) {
+            cp0_regs[CP0_CAUSE_REG] |= CP0_CAUSE_BD;
+        }
+        else {
+            cp0_regs[CP0_CAUSE_REG] &= ~CP0_CAUSE_BD;
+        }
+    }
+    else
+    {
+        if (r4300->emumode != EMUMODE_PURE_INTERPRETER)
+        {
+            cp0_regs[CP0_EPC_REG] = (w != 2)
+                ? *r4300_pc(r4300)
+                : address;
+        }
+        else {
+            cp0_regs[CP0_EPC_REG] = *r4300_pc(r4300);
+        }
+
+        cp0_regs[CP0_CAUSE_REG] &= ~CP0_CAUSE_BD;
+        cp0_regs[CP0_STATUS_REG] |= CP0_STATUS_EXL;
+
+        if (address >= UINT32_C(0x80000000) && address < UINT32_C(0xc0000000)) {
+            usual_handler = 1;
+        }
+
+        for (i = 0; i < 32; i++)
+        {
+            if (/*r4300->cp0.tlb.entries[i].v_even &&*/ address >= r4300->cp0.tlb.entries[i].start_even &&
+                    address <= r4300->cp0.tlb.entries[i].end_even) {
+                usual_handler = 1;
+            }
+            if (/*r4300->cp0.tlb.entries[i].v_odd &&*/ address >= r4300->cp0.tlb.entries[i].start_odd &&
+                    address <= r4300->cp0.tlb.entries[i].end_odd) {
+                usual_handler = 1;
+            }
+        }
+
+        generic_jump_to(r4300, (usual_handler)
+                ? UINT32_C(0x80000180)
+                : UINT32_C(0x80000000));
+    }
+
+    if (r4300->delay_slot == 1 || r4300->delay_slot == 3)
+    {
+        cp0_regs[CP0_CAUSE_REG] |= CP0_CAUSE_BD;
+        cp0_regs[CP0_EPC_REG] -= 4;
+    }
+    else
+    {
+        cp0_regs[CP0_CAUSE_REG] &= ~CP0_CAUSE_BD;
+    }
+    if (w != 2) {
+        cp0_regs[CP0_EPC_REG] -= 4;
+    }
+
+    r4300->cp0.last_addr = *r4300_pc(r4300);
+
+    if (r4300->emumode == EMUMODE_DYNAREC)
+    {
+        dyna_jump();
+        if (!r4300->dyna_interp) { r4300->delay_slot = 0; }
+    }
+
+    if (r4300->emumode != EMUMODE_DYNAREC || r4300->dyna_interp)
+    {
+        r4300->dyna_interp = 0;
+        if (r4300->delay_slot)
+        {
+            r4300->skip_jump = *r4300_pc(r4300);
+            *r4300_cp0_next_interrupt(&r4300->cp0) = 0;
+        }
+    }
+}
+
+void exception_general(struct r4300_core* r4300)
+{
+    uint32_t* cp0_regs = r4300_cp0_regs(&r4300->cp0);
+
+    cp0_update_count(r4300);
+    cp0_regs[CP0_STATUS_REG] |= CP0_STATUS_EXL;
+
+    cp0_regs[CP0_EPC_REG] = *r4300_pc(r4300);
+
+    if (r4300->delay_slot == 1 || r4300->delay_slot == 3)
+    {
+        cp0_regs[CP0_CAUSE_REG] |= CP0_CAUSE_BD;
+        cp0_regs[CP0_EPC_REG] -= 4;
+    }
+    else
+    {
+        cp0_regs[CP0_CAUSE_REG] &= ~CP0_CAUSE_BD;
+    }
+
+    generic_jump_to(r4300, UINT32_C(0x80000180));
+
+    r4300->cp0.last_addr = *r4300_pc(r4300);
+
+    if (r4300->emumode == EMUMODE_DYNAREC)
+    {
+        dyna_jump();
+        if (!r4300->dyna_interp) { r4300->delay_slot = 0; }
+    }
+    if (r4300->emumode != EMUMODE_DYNAREC || r4300->dyna_interp)
+    {
+        r4300->dyna_interp = 0;
+        if (r4300->delay_slot)
+        {
+            r4300->skip_jump = *r4300_pc(r4300);
+            *r4300_cp0_next_interrupt(&r4300->cp0) = 0;
+        }
+    }
 }
