+++ src/device/r4300/cp0.h
@@ -211,5 +211,8 @@ int check_cop1_unusable(struct r4300_core* r4300);
 
 void cp0_update_count(struct r4300_core* r4300);
 
+void TLB_refill_exception(struct r4300_core* r4300, uint32_t address, int w);
+void exception_general(struct r4300_core* r4300);
+
 #endif /* M64P_DEVICE_R4300_CP0_H */
