+++ src/device/r4300/interrupt.c
@@ -37,7 +37,7 @@
 #include "api/m64p_types.h"
 #include "device/pif/bootrom_hle.h"
 #include "device/r4300/cached_interp.h"
-#include "device/r4300/exception.h"
+#include "device/r4300/cp0.h"
 #include "device/r4300/new_dynarec/new_dynarec.h"
 #include "device/r4300/r4300_core.h"
 #include "device/r4300/recomp.h"
