+++ src/device/r4300/new_dynarec/new_dynarec.c
@@ -34,13 +34,13 @@
 #include "main/rom.h"
 #include "device/memory/memory.h"
 #include "device/r4300/cached_interp.h"
+#include "device/r4300/cp0.h"
 #include "device/r4300/cp1.h"
 #include "device/r4300/interrupt.h"
 #include "device/r4300/ops.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/tlb.h"
 #include "device/r4300/fpu.h"
-#include "device/r4300/exception.h"
 #include "device/rcp/mi/mi_controller.h"
 #include "device/rcp/rsp/rsp_core.h"
