+++ src/device/r4300/tlb.c
@@ -22,7 +22,6 @@
 #include "tlb.h"
 
 #include "api/m64p_types.h"
-#include "device/r4300/exception.h"
 #include "device/r4300/r4300_core.h"
 #include "device/memory/memory.h"
 #include "main/rom.h"
