+++ src/device/r4300/x86/dynarec.c
@@ -30,7 +30,6 @@
 #include "device/r4300/cached_interp.h"
 #include "device/r4300/cp0.h"
 #include "device/r4300/cp1.h"
-#include "device/r4300/exception.h"
 #include "device/r4300/interrupt.h"
 #include "device/r4300/macros.h"
 #include "device/r4300/ops.h"
